fix: split general tests by stack Created new tests files for each stack on the new folder stacks
Resolves: #142
